+++ extensions/hello/extension.php
@@ -12,9 +12,16 @@ return [
 
     'controllers' => 'src/Controller/*Controller.php',
 
-    'settings' => [
+    'parameters' => [
 
-        'system'  => 'extension://hello/views/admin/settings.razr'
+        'settings' => [
+
+            'view' => 'extension://hello/views/admin/settings.razr',
+            'defaults' => [
+                'message' => 'World'
+            ]
+
+        ]
 
     ],