+++ themes/alpha/theme.php
@@ -1,16 +1,30 @@
-
 <?php
 
+/**
+ * Theme configuration file.
+ *
+ * @link http://pagekit.com/docs/themes - basic explanation of theme development
+ * @link http://pagekit.com/docs/configuration - full documentation of config options
+ */
 return [
 
+    /**
+     * Namespace to autoload theme classes.
+     */
     'autoload' => [
 
         'Pagekit\\Alpha\\' => 'src'
 
     ],
 
+    /**
+     * The main theme class to be loaded when the theme is booted.
+     */
     'main' => 'Pagekit\\Alpha\\AlphaTheme',
 
+    /**
+     * Overwrite default template files with templates provided by the theme and define stream wrappers for shorter path access.
+     */
     'resources' => [
 
         'override' => [
@@ -19,13 +33,24 @@ return [
 
     ],
 
-    'settings' => [
+    /**
+     * Define default settings values and views where end users can change these values.
+     */
+    'parameters' => [
 
-        'system'  => 'theme://alpha/views/admin/settings.razr',
-        'widgets' => 'theme://alpha/views/admin/widgets/edit.razr'
+        'settings' => [
+            'view' => 'theme://alpha/views/admin/settings.razr'
+        ],
+
+        'widgets'  => [
+            'view' => 'theme://alpha/views/admin/widgets/edit.razr'
+        ]
 
     ],
 
+    /**
+     * Widget positions offered by this theme. These positions will be rendered in different locations of the theme's template.
+     */
     'positions' => [
 
         'logo'       => 'Logo',
@@ -39,6 +64,9 @@ return [
 
     ],
 
+    /**
+     * List of renderers provided by this theme. A renderer determines the markup to be used in a widget position.
+     */
     'renderer' => [
 
         'blank'     => 'theme://alpha/views/renderer/position.blank.razr',